+++ frontend/src/components/article.js
@@ -8,7 +8,7 @@ export default class Article extends Component {
     this.state = {
       article: "",
       title: "",
-      author: "Anonymous",
+      author: "",
       type: "",
       key: "",
       journal: "",
@@ -27,40 +27,8 @@ export default class Article extends Component {
     axios
       .get("/entries/id" + this.props.match.params.id)
       .then((response) => {
-        // let myString = JSON.stringify(response.data, "", " ")
-        //   .trim()
-        //   .replace("{", "")
-        //   .replace("}", "");
-
-        //console.log(myString);
-        // console.log(response.data.title);
-        // console.log(response.data.author);
-        // console.log(response.data.type);
-        // console.log(response.data.key);
-        // console.log(response.data.journal);
-        // console.log(response.data.pages);
-        // console.log(response.data.volume);
-        // console.log(response.data.annote);
-        // console.log(response.data.publisher);
-        // console.log(response.data.method);
-        // console.log(response.data.participants);
-        // console.log(response.data.year);
-        // console.log(response.data.month);
-
+        console.log(response.data);
         this.setState(response.data);
-        // this.setState({ title: response.data.title });
-        // this.setState({ author: response.data.author });
-        // this.setState({ type: response.data.type });
-        // this.setState({ key: response.data.key });
-        // this.setState({ journal: response.data.journal });
-        // this.setState({ pages: response.data.pages });
-        // this.setState({ volume: response.data.volume });
-        // this.setState({ annote: response.data.annote });
-        // this.setState({ publisher: response.data.publisher });
-        // this.setState({ method: response.data.method });
-        // this.setState({ participants: response.data.participants });
-        // this.setState({ year: response.data.year });
-        // this.setState({ month: response.data.month });
       })
       .catch(function (error) {
         console.log(error);
@@ -73,46 +41,48 @@ export default class Article extends Component {
         <p>{this.state.article}</p>
         <h3>{this.state.title} </h3>
         <h6>{this.state.author} </h6>
-        <p>
+        <hr />
+        <p style={{ display: this.state.type ? "" : "none" }}>
           <b>Type: </b>
           {this.state.type}
         </p>
-        <p>
+        <p style={{ display: this.state.journal ? "" : "none" }}>
           <b>Journal: </b>
           {this.state.journal}
         </p>
-        <p>
+        <p style={{ display: this.state.pages ? "" : "none" }}>
           <b>Pages: </b>
           {this.state.pages}
         </p>
-        <p style={{ visibility: this.state.volume ? "visible" : "collapse" }}>
+        <p style={{ display: this.state.volume ? "" : "none" }}>
           <b>volume: </b>
           {this.state.volume}
         </p>
-        <p>
+        <p style={{ display: this.state.annote ? "" : "none" }}>
           <b>Annote: </b>
           {this.state.annote}
         </p>
-        <p>
+        <p style={{ display: this.state.publisher ? "" : "none" }}>
           <b>Publisher: </b>
           {this.state.publisher}
         </p>
-        <p>
+        <p style={{ display: this.state.method ? "" : "none" }}>
           <b>Method: </b>
           {this.state.method}
         </p>
-        <p>
+        <p style={{ display: this.state.participants ? "" : "none" }}>
           <b>Participants: </b>
           {this.state.participants}
         </p>
-        <p>
+        <p style={{ display: this.state.year ? "" : "none" }}>
           <b>Year: </b>
           {this.state.year}
         </p>
-        <p>
+        <p style={{ display: this.state.month ? "" : "none" }}>
           <b>Month: </b>
           {this.state.month}
         </p>
+        <hr />
       </div>
     );
   }
